refactor: 패키지 수정
- 프로필과 게시글 이미지 경로 변경
