+++ backend/src/main/java/edonymyeon/backend/image/profileimage/repository/ProfileImageInfoRepository.java
@@ -1,6 +1,6 @@
-package edonymyeon.backend.image.postimage.repository;
+package edonymyeon.backend.image.profileimage.repository;
 
-import edonymyeon.backend.image.postimage.ProfileImageInfo;
+import edonymyeon.backend.image.profileimage.domain.ProfileImageInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProfileImageInfoRepository extends JpaRepository<ProfileImageInfo, Long> {
