+++ backend/src/test/java/edonymyeon/backend/support/ProfileImageInfoTestSupport.java
@@ -1,7 +1,7 @@
 package edonymyeon.backend.support;
 
-import edonymyeon.backend.image.postimage.ProfileImageInfo;
-import edonymyeon.backend.image.postimage.repository.ProfileImageInfoRepository;
+import edonymyeon.backend.image.profileimage.domain.ProfileImageInfo;
+import edonymyeon.backend.image.profileimage.repository.ProfileImageInfoRepository;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
